% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_cleaning.R
\name{eq_location_cleaning}
\alias{eq_location_cleaning}
\title{Location cleaning}
\usage{
eq_location_cleaning(raw_data)
}
\description{
Location cleaning
}
\details{
This function cleanes the location variable
}
\examples{
raw_data <- eq_read_data()
cleaned_data <- eq_location_cleaning(raw_data)

}
