% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_data.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Read data}
\usage{
eq_read_data()
}
\description{
Read data
}
\details{
This function opens the earthquakes dataset.
}
\examples{
raw_data <- eq_read_data()

}
