% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Earthquake label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Cleaned data from eq_clean_data function.}
}
\description{
Earthquake label
}
\details{
This function creates earthquake map label.
}
\examples{
raw_data <- eq_read_data()
clean_data <- eq_clean_data(raw_data)
clean_data \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")

}
