% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map_label.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake map}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{Cleaned data from eq_clean_data function.}

\item{annot_col}{Annotation column.}
}
\description{
Earthquake map
}
\details{
This function creates earthquake map.
}
