% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Geom timeline label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Mapping}

\item{data}{Data}

\item{stat}{Stat}

\item{position}{Position}

\item{na.rm}{NA remove}

\item{show.legend}{Show legend}

\item{inherit.aes}{Inherit aes}
}
\description{
Geom timeline label
}
\details{
This function add label to the created custom timeline chart.
}
